Validate root-level devictree entries in the keymap devicetree
In order to prevent unsafe devicetree overriding via keymap files, require the
root level devicetree entries in the provided keymap to match a strict
allow-list. This is in fact more restrictive than necessary, because it'd be
legitimate to define for example a `compatible = "zmk,behavior-hold-tap";`
section at the top level, but it's much easier to safely validate.